+++ app/Http/Controllers/Admin/SubscriptionCategoryController.php
@@ -9,16 +9,22 @@ use Illuminate\Http\Request;
 class SubscriptionCategoryController extends Controller
 {
     public function index() {
+        $this->authorize('viewAny', SubscriptionCategory::class);
+
         $subscription_categories = SubscriptionCategory::query()->paginate(30);
         return view('admin.subscriptions.categories', compact('subscription_categories'));
     }
 
     public function store(Request $request) {
+        $this->authorize('create', SubscriptionCategory::class);
+
         (new SubscriptionCategory($request->all()))->save();
         return redirect()->back()->with('success', 'Категория успешно создана');
     }
 
     public function update(Request $request) {
+        $this->authorize('update', SubscriptionCategory::class);
+
         $subscription = SubscriptionCategory::query()->find($request->id);
         if (!$subscription) {
             return redirect()->back()->with('error', 'Категория не найдена');
@@ -34,6 +40,8 @@ class SubscriptionCategoryController extends Controller
     }
 
     public function show(Request $request) {
+        $this->authorize('view', SubscriptionCategory::class);
+
         $subscription = SubscriptionCategory::query()->find($request->id);
         if (!$subscription) {
             return response()->json([
@@ -45,6 +53,8 @@ class SubscriptionCategoryController extends Controller
     }
 
     public function delete(Request $request) {
+        $this->authorize('delete', SubscriptionCategory::class);
+
         $subscription = SubscriptionCategory::query()->find($request->id);
         if($subscription) {
             $subscription->delete();
